refactor(ipod): split useIpodStore into slice modules
